+++ CHANGELOG.md
@@ -9,9 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* `left` field in `IosSafeArea`
-* `right` field in `IosSafeArea`
-
+* support `left`&`right` safearea [#9](https://github.com/rustunit/bevy_ios_safearea/pull/9)
+ 
 ## [0.1.6] - 2025-01-15
 
 ### Fixed